==> 0.Finetune_folder_structure_for_github_use.R <==
#Finetune_folder_structure_for_github_use
if (file.exists("./data/internal")==FALSE){dir.create("./data/internal")}
if (file.exists("./data/internal/telemetry")==FALSE){dir.create("./data/internal/telemetry")}
if (file.exists("./data/internal/env")==FALSE){dir.create("./data/internal/env")}
if (file.exists("./data/internal/SEM")==FALSE){dir.create("./data/internal/SEM")}
if (file.exists("./data/internal/power_data_agg_fixed_receivers")==FALSE){dir.create("./data/internal/power_data_agg_fixed_receivers")}
if (file.exists("./data/internal/power_data_agg_random_receivers")==FALSE){dir.create("./data/internal/power_data_agg_random_receivers")}
if (file.exists("./data/internal/power_databin_random_receivers")==FALSE){dir.create("./data/internal/power_databin_random_receivers")}
if (file.exists("./data/internal/power_data_agg_all_receiver_combinations")==FALSE){dir.create("./data/internal/power_data_agg_all_receiver_combinations")}
if (file.exists("./models")==FALSE){dir.create("./models")}
if (file.exists("./models/single_wavelet_data_agg")==FALSE){dir.create("./models/single_wavelet_data_agg")}
if (file.exists("./models/single_wavelet_data_agg_tilt")==FALSE){dir.create("./models/single_wavelet_data_agg_tilt")}
if (file.exists("./models/single_wavelet_data_agg_noise")==FALSE){dir.create("./models/single_wavelet_data_agg_noise")}
if (file.exists("./models/single_wavelet_data_agg_water_height")==FALSE){dir.create("./models/single_wavelet_data_agg_water_height")}
if (file.exists("./models/single_wavelet_data_agg_temp")==FALSE){dir.create("./models/single_wavelet_data_agg_temp")}
if (file.exists("./models/single_wavelet_data_agg_cond")==FALSE){dir.create("./models/single_wavelet_data_agg_cond")}
if (file.exists("./models/single_wavelet_data_agg_wind")==FALSE){dir.create("./models/single_wavelet_data_agg_wind")}
if (file.exists("./models/single_wavelet_data_agg_q")==FALSE){dir.create("./models/single_wavelet_data_agg_q")}
if (file.exists("./models/single_wavelet_data_agg_q_abs")==FALSE){dir.create("./models/single_wavelet_data_agg_q_abs")}
if (file.exists("./models/single_wavelet_data_agg_q_magn")==FALSE){dir.create("./models/single_wavelet_data_agg_q_magn")}
if (file.exists("./models/loess_tilt_angle_models")==FALSE){dir.create("./models/loess_tilt_angle_models")}
if (file.exists("./models/loess_noise_models")==FALSE){dir.create("./models/loess_noise_models")}
if (file.exists("./models/coherence_wavelet_data_agg_water_height")==FALSE){dir.create("./models/coherence_wavelet_data_agg_water_height")}
if (file.exists("./models/coherence_wavelet_data_agg_q_abs")==FALSE){dir.create("./models/coherence_wavelet_data_agg_q_abs")}
if (file.exists("./models/coherence_wavelet_data_agg_q")==FALSE){dir.create("./models/coherence_wavelet_data_agg_q")}
if (file.exists("./models/coherence_wavelet_data_agg_noise")==FALSE){dir.create("./models/coherence_wavelet_data_agg_noise")}
if (file.exists("./figures")==FALSE){dir.create("./figures")}
if (file.exists("./models/SEM")==FALSE){dir.create("./models/SEM")}
if (file.exists("./models/GLM")==FALSE){dir.create("./models/GLM")}
